pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "predict-structure"
version = "0.1.0"
description = "Unified protein structure prediction CLI for BV-BRC (Boltz-2, Chai-1, AlphaFold2, ESMFold)"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Andreas Wilke"}
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
keywords = ["protein", "structure prediction", "BV-BRC", "Boltz", "Chai", "AlphaFold", "ESMFold"]
dependencies = [
    "click>=8.0",
    "biopython>=1.80",
    "numpy>=1.23",
    "pandas>=1.5",
]

[project.optional-dependencies]
boltz = [
    "pyyaml>=6.0",
]
chai = [
    "pyarrow>=12.0",
]
esmfold = [
    "torch>=2.0",
    "transformers>=4.30",
    "accelerate>=0.20",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "black>=23.0",
    "ruff>=0.1",
    "mypy>=1.0",
]
all = [
    "predict-structure[boltz,chai,esmfold,dev]",
]

[project.scripts]
predict-structure = "predict_structure.cli:main"

[project.urls]
Repository = "https://github.com/CEPI-dxkb/PredictStructureApp"

[tool.setuptools.packages.find]
where = ["."]
include = ["predict_structure*"]

[tool.black]
line-length = 99
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 99
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
addopts = "-v --tb=short"
